Make inclusion of railtie opt-in
